+++ scribo_ct/examples/example_1/source/main.c
@@ -3,22 +3,26 @@
 // File: main.c
 // ---------------------------------------------------------------------------------------------------------------------
 
+// Produce log messages and invoke functions from other compilation units
+
 #define SCRIBO_CATEGORY APP
 #include <scribo.h>
 #include "foo.h"
 #include "bar.h"
 #include "baz.h"
+#include "qux.h"
 
 int main(int argc, char* argv[])
 {
     if (argc >= 1)
     {
-        SCRIBO(LOG, "Running %s", argv[0]);                                              // Category APP, verbosity LOG.
+        SCRIBO(LOG, "Running %s", argv[0]);                                               // Category APP, verbosity LOG
     }
     doFoo();
     doBar();
     doBaz();
-    SCRIBO(LOG, "The end");                                                              // Category APP, verbosity LOG.
+    doQux();
+    SCRIBO(LOG, "The end");                                                               // Category APP, verbosity LOG
     return 0;
 }